requirements.txt
# Economic Data Platform - Unified Requirements
# Core database and async support
sqlalchemy==2.0.36
psycopg[binary]==3.2.9
asyncpg==0.29.0
alembic==1.12.1
greenlet==3.1.1

# API clients and HTTP
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Data processing and analysis
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Financial data APIs
yfinance==0.2.65
fredapi==0.5.1

# News API integration
newsapi-python==0.2.7

# Time series and correlation analysis
statsmodels==0.14.0
scikit-learn==1.3.2

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Async utilities and background tasks
asyncio-throttle==1.0.2
tenacity==8.2.3

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Configuration and environment
python-dotenv==1.0.0
click==8.1.7

# UUID and JSON handling
uuid==1.30

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Vector database integration
chromadb==0.4.24
sentence-transformers==2.7.0
transformers==4.44.2
huggingface_hub==0.24.6

# Optional: Web framework for API endpoints
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6